+++ changelog.md
@@ -19,3 +19,4 @@
 - Add Service Name ENV Variable to Zookeeper and Kafka to match Consul DNS
 - Add Call_Suppress to tell mesos it does not send us offers until we ask
 - Add default values for some env
+- Add custom image name via env